+++ cypress/cypress/component/Datepicker/DatePickerInput.cy.tsx
@@ -1,14 +1,15 @@
 /* eslint-disable max-lines */
-import { DatePickerInput, Form } from "react-hook-form-components";
+import { DatePickerInput, Form, setUtcTimeToZero } from "react-hook-form-components";
 import "react-datepicker/dist/react-datepicker.css";
 import { faker } from "@faker-js/faker";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
 
-it("basic example works", () => {
+it("selecting today works", () => {
   const name = faker.random.alpha(10);
+
   const schema = yup.object().shape({
-    [name]: yup.string().required(),
+    [name]: yup.date().required(),
   });
 
   cy.mount(
@@ -24,10 +25,75 @@ it("basic example works", () => {
   cy.get("input[type=submit]").click({ force: true });
 
   const todayMidnight = new Date();
-  todayMidnight.setHours(0, 0, 0, 0);
-  todayMidnight.setMinutes(todayMidnight.getMinutes() - todayMidnight.getTimezoneOffset());
+  setUtcTimeToZero(todayMidnight);
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: todayMidnight });
+});
 
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: todayMidnight.toString() });
+it("setting intial value as iso string works", () => {
+  const name = faker.random.alpha(10);
+  const randomDate = faker.date.future();
+
+  const schema = yup.object().shape({
+    [name]: yup.date().required(),
+  });
+
+  const x = (values: unknown) => {
+    console.log(values);
+    console.log(JSON.stringify(values));
+  };
+
+  cy.mount(
+    <Form
+      onSubmit={cy.spy(x).as("onSubmitSpy")}
+      resolver={yupResolver(schema)}
+      defaultValues={{
+        [name]: randomDate.toISOString(),
+      }}
+    >
+      <DatePickerInput name={name} label={name} />
+
+      <input type={"submit"} />
+    </Form>,
+  );
+
+  cy.get("input[type=submit]").click({ force: true });
+
+  setUtcTimeToZero(randomDate);
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: randomDate });
+});
+
+it("setting intial value as date object works", () => {
+  const name = faker.random.alpha(10);
+  const randomDate = faker.date.future();
+
+  const schema = yup.object().shape({
+    [name]: yup.date().required(),
+  });
+
+  cy.mount(
+    <Form
+      onSubmit={cy.spy().as("onSubmitSpy")}
+      resolver={yupResolver(schema)}
+      defaultValues={{
+        [name]: randomDate,
+      }}
+    >
+      <DatePickerInput name={name} label={name} />
+
+      <input type={"submit"} />
+    </Form>,
+  );
+
+  cy.get("input[type=submit]").click({ force: true });
+
+  setUtcTimeToZero(randomDate);
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: randomDate });
 });
 
 it("is disabled", () => {