+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for iso strings as default values for the `Form` component
+
 ### Changed
 
+- Fixed `DatePickerInput` issues
 - Dependabot: Bump http-cache-semantics from 4.1.0 to 4.1.1
 
 ## [0.5.0] - 2023-01-25